Revert "Fetch URL for likes added"
This reverts commit 5ea28fcb4d4c3def7529c308773c9f871224ce73.
