+++ code/src/components/Thought.js
@@ -8,27 +8,13 @@ export const Thought = ({
   id,
   onLikedThought
 }) => {
-  const [likedThought, setLikedThougt] = useState([]);
-
-  const getLikedThought = (id) => {
-    const updatedThoughts = thoughts.map((thought) => {
-      if (id === thought._id) {
-        thought.hearts += 1;
-      }
-      return thought;
-    });
-    setThoughts(updatedThoughts);
-    if (!likedThought.includes(id)) {
-      setLikedThougt([...likedThought, id]);
-    }
-  };
   return (
     <div className="thought-container">
       <p className="thought" id={id}>
         {message}
       </p>
       <div className="liked-container">
-        <ThoughtLikes id={id} onLikedThought={() => onLikedThought(id)} />
+        <ThoughtLikes id={id} onLikedThought={onLikedThought} />
         <p className="liked"> x {hearts}</p>
       </div>
       <p className="thought-time">{time}</p>
