+++ code/src/components/ThoughtList.js
@@ -4,7 +4,28 @@ import { Thought } from './Thought'
 import { THOUGHTS_URL } from '../urls';
 
 export const ThoughtList = ({ thoughts, setThoughts }) => {
+  const [likedThought, setLikedThougt] = useState([]);
 
+  const getLikedThought = (id) => {
+    const updatedThoughts = thoughts.map((thought) => {
+      if (id === thought._id) {
+        thought.hearts += 1;
+      }
+      return thought;
+    });
+    setThoughts(updatedThoughts);
+    if (!likedThought.includes(id)) {
+      setLikedThougt([...likedThought, id]);
+    }
+  };
+
+  const onLikedThought = (id) => {
+    fetch(`${THOUGHTS_URL}/${id}/like`, {
+      method: 'POST',
+      body: '',
+      headers: { 'Content-Type': 'application/json' }
+    }).then(() => getLikedThought(id));
+  };
 
   return (
     <div>
