+++ code/src/urls.js
@@ -1,1 +1,1 @@
-export const THOUGHTS_URL = 'https://happy-thoughts-technigo.herokuapp.com/thoughts'
+export const THOUGHTS_URL = 'https://happiest-happy-thoughts.herokuapp.com/thoughts'
